Move all fetching code to `_app.tsx`
