+++ app/pages/index.tsx
@@ -1,114 +1,14 @@
-import * as anchor from '@project-serum/anchor';
 import { NextPage } from 'next';
 import { useRouter } from 'next/router';
-import { ConfirmOptions } from '@solana/web3.js';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback } from 'react';
 import { WalletMultiButton } from '@solana/wallet-adapter-react-ui';
-import { AnchorWallet, useAnchorWallet } from '@solana/wallet-adapter-react';
-
-import IDL from '../../target/idl/nft_club.json';
-import { NftClub } from '../../target/types/nft_club';
+import { useAnchorWallet } from '@solana/wallet-adapter-react';
 import { useUser } from '../hooks/userUser';
-
-const PROGRAM_ID = new anchor.web3.PublicKey(
-  '6dND1tHXuvCzB9Fe88FvnrZEqTVraPWGxtR5HQs4Z3dx'
-);
-const OPTS = {
-  preflightCommitment: 'processed',
-} as ConfirmOptions;
-
-const endpoint = 'https://api.devnet.solana.com';
-const connection = new anchor.web3.Connection(
-  endpoint,
-  OPTS.preflightCommitment
-);
 
 const Home: NextPage = () => {
   const router = useRouter();
   const connectedWallet = useAnchorWallet();
-
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const { user, setUser } = useUser();
-
-  const program = useMemo(() => {
-    if (connectedWallet) {
-      const provider = new anchor.Provider(connection, connectedWallet, OPTS);
-
-      return new anchor.Program<NftClub>(
-        IDL as unknown as NftClub,
-        PROGRAM_ID,
-        provider
-      );
-    }
-
-    return null;
-  }, [connectedWallet]);
-
-  const getCreatorAccountForUserWallet = useCallback(
-    async (
-      nftClubProgram: anchor.Program<NftClub>,
-      wallet: AnchorWallet
-    ): Promise<Record<string, unknown>> => {
-      const creatorSeeds = [
-        wallet.publicKey.toBuffer(),
-        anchor.utils.bytes.utf8.encode('creator'),
-      ];
-
-      const [creatorPubKey] = await anchor.web3.PublicKey.findProgramAddress(
-        creatorSeeds,
-        nftClubProgram.programId
-      );
-
-      return await nftClubProgram.account.creator.fetch(creatorPubKey);
-    },
-    []
-  );
-
-  const fetchSubscriptionsForUserWallet = useCallback(
-    async (
-      nftClubProgram: anchor.Program<NftClub>,
-      wallet: AnchorWallet
-    ): Promise<Record<string, unknown>[]> => {
-      return await nftClubProgram.account.subscription.all([
-        {
-          memcmp: {
-            offset: 8,
-            bytes: wallet.publicKey.toBase58(),
-          },
-        },
-      ]);
-    },
-    []
-  );
-
-  const fetchUserDetails = useCallback(
-    async (nftClubProgram: anchor.Program<NftClub>, wallet: AnchorWallet) => {
-      const creator = await getCreatorAccountForUserWallet(
-        nftClubProgram,
-        wallet
-      );
-      const subscriptions = await fetchSubscriptionsForUserWallet(
-        nftClubProgram,
-        wallet
-      );
-
-      (creator || subscriptions.length) &&
-        setUser({
-          subscriptions,
-          creatorAccount: creator,
-        });
-
-      setIsLoading(false);
-    },
-    [getCreatorAccountForUserWallet, fetchSubscriptionsForUserWallet, setUser]
-  );
-
-  useEffect(() => {
-    if (connectedWallet && program) {
-      setIsLoading(true);
-      fetchUserDetails(program, connectedWallet);
-    }
-  }, [connectedWallet, program, fetchUserDetails]);
+  const { user } = useUser();
 
   const handleBecomeCreator = useCallback(() => {
     router.push('/sign-up');
@@ -163,7 +63,7 @@ const Home: NextPage = () => {
               onClick={() => {
                 user.subscriptions.length
                   ? router.push('/subscription-hub')
-                  : router.push('/subcribe');
+                  : router.push('/creator-landing-page');
               }}
             >
               {user.subscriptions.length
